Consolidate parser state declaration
